Tagging release 9.12.
This is being done after the fact.  After my computer was stolen I
recovered the svn repo from a backup.  I'm not sure whether anything
was lost, but I don't see a tag for this release.  The current repo
state appears to match what was released (at least the main files
match).