Adding new option --no-string.  This disables use of std::string wrapping for string outputs, and uses (char*, int) argument pairs instead.
This was motivated by some problems with Python-Qt and Java wrapper
code, which would both crash when calling into string output
conversion functions.  Seemed to be due to a C++ library mismatch
causing problems with std::string.

In fortwrap.py, this option sets opts.std_string to False.  The
character array size argument is added to the wrapper function, but is
not internally added to the argument list, since that would require
more work for argument ordering/numbering.  Instead, the extra
argument is just inserted at runtime when writing the function
definitions.

To try and prevent confusion, the extra argument is given the suffix
'_size__', to distinguish it from the hidden length argument sent to
Fortran, which has suffix '_len__'.

Currently no checking is done against the Fortran string size (when
known), although I supposed something could be added.  For example if
Fortran will complain if a larger string is passed in, the wrapper
code could truncate by using min() to determine the value of the
hidden length argument.

It occurs to me know that a potential drawback here is that now the
user has to know how much character data to allocate before calling
the function, in the case of Fortran functions with specified
character length.  A third option would be to create a wrapper class
(a simpler version of std::string) that can do the allocation
automatically.