Adding support for wrapping of type bound procedures with NOPASS
Corresponding static methods are created in C++.  Currently there is a
limitation in that association of procedures with NOPASS TBP's is done
in a global namespace.  If the target name of the NOPASS TBP is not
unique (i.e., multiple procedures in the project have that same name),
FortWrap will not necessarily associate the correct procedure, since
it is not fully aware of the compilation namespace introduced through
USE declarations.