Parse PROCEDURE arguments
Update parsing to recognize arguments that are PROCEDURE, but do not
have POINTER attribute (previously would print an error message).
These are marked as non-interoperable and not wrapped (at least from
initial testing, the current wrapping approach does not seem to work
for these cases, only when the POINTER attribute is used).