Changing behavior for dtor wrapping
New behavior is that by default, Fortran dtors are not included as
separate functions in the C++ wrapper code (they are called
automatically by the C++ destructor, as before).  Use %include to
include the dtor in the API as a separate function.  In either case,
the Fortran dtor is still automatically called by the C++ destructor,
and multiple calls are protected via an internal state variable.

Any previous configuration files that used %ignore to hide dtors from
the API will need to be updated, because %ignore'ing a dtor will now
cause it to not be wrapped at all (and not called by the C++
destructor).

Also adding some additional features to the derived_types test case.
Adding a private derived type to verify that it doesn't get wrapped or
cause errors, and adding a destructor.