Adding DLLEXPORT preprocessor macro
Adding preprocessor macro that adds the __declspec(dllexport)
attribute to class definitions.  This is needed when compiling the
generated wrapper code with the MSC compiler.

Although FortWrap currently targets wrapper code that interfaces with
Fortran code compiled using gfortran, there are cases where one may
want to compile the wrapper code with MSC.  One example is when
creating an extension module for Python 3 on Windows, in which case
the MSC compiler may be necessary for compiling the Python bindings
and C++ code, whereas the underlying Fortran code could still be
compiled with gfortran.