Work in progess commit
Have some of the wrapper generation code working for derived type
pointer return values.  With some manual editing can get the test to
compile.  However, the problem is how to construct an instance of the
object to be returned?  Key is to not call the Fortran "allocate"
function, since we want to manually set up the data_ptr instead of
having Fortran allocate it.  One way of doing this would be to create
a fake constructor that doesn't call the allocator; drawback is that
adds to the API (may be able to avoid that by making the constructor
private and making classes that need it friend classes)