More work in progress
Write a special pointer constructor that can be used to generate
objects based on a pointer.  By manually editing the cpp file to use
this pointer constructor in the pointer return value location, the
code compiles.  Also added the "owns" attribute to protect calls to
the Fortran deallocator (however, when I tested removing this check,
the code still ran without crashing...)