Test Fortran ctors/dtors with pointer return handling
Updated the acquire function so that it sets initialized to true.
This means manually calling acuire will trigger the C++ destructor to
call the Fortran dtor, if any.  This might need further review;
potentially the ownership API could also allow adjustment of the
initialized flag as well to provide more control.