Adding class_data_ptr member for class wrapping
Set this up to point to the address of class_data.  Updated wrapping
code to pass this instead of the address of class_data.  This is in
preparation for trying to wrap class pointer return values, which
hopefully can just set the value of this pointer.  This update
also simplifies the code generation, since now argument passing is
more consistent with that for TYPE data.