Adding additional tests for class pointer return
Appears that for CLASS(), POINTER return values, gfortran just returns
the address of the derived type data, not the address of the class
container.  These needs to be investigated further, as it's unclear
how the dynamic type is stored.  Current wrapping approach seems to
work for wrapping pointer return for either TYPE or CLASS, but more
testing needed.