Initial take at array wrapping
Adding utility functions for the Argument class to write Fortran
wrapper local variable declarations as well as "setup code" that
happens prior to the function call.  The "arrays" test now works minus
the subroutine (wrapping only set up for functions so far) and minus
the assumed size version (which may not be practical to wrap with the
ISO_C_BINDING approach).