Add notes about procedure pointer intent
Wrapping approach passes the procedure pointer by value.  But
wrapping only intent(in) does not seem to be suitable because that
seems to disallow some useful cases where the function being wrapped
stores a copy of the function pointer (unclear why the compiler does
not allow intent(in) in these cases...)  So as with previous code,
intent(out) is not wrapped, but intent(in) is not required.