Don't wrap procedures that are not in a module
Without additional declarations, the ISO_C_BINDING wrapping code won't
compile because it needs an interface to call the original procedure.

Updating the c_binding_name to use "global_" for these cases, just in
case decide to wrap them in the future.

Add notes to no_module test that it is not supported.  And
specifically add it to excludes in run_tests.