Minor tweak and classes tests now run
Both classes and classes_debug tests now run.

The current class container wrapping approach appears to work fine
with gfortran.  However, it may not be completely portable because
class container pointers are used interchangeably within a particular
inheritance structure.  E.g. a pointer to a container around a child
instance is passed where a pointer to a base container is expected.

A more portable approach could use one container for a given class
hierarchy.  Each allocater function would allocate the appropriate
type.  The trick with this is that "select type" statements would be
needed in the wrapper code when trying to call methods that are not
present in the base class.