Rework class container approach
Don't create class containers for child classes, only the base
classes.  Add SELECT TYPE code around wrapper functions when need to
access functionality that is specific to the child class.

This is in effort to make the wrapping more portable.  Previously
class container pointers were mixed so that the actual container type
could be different than what the compiler was expecting.  New approach
tries to be more rigorous with this.

Tests pass, but further review needed.