Rework handling of Fortran-only arguments
Rework so that Fortran-only arguments do not appear in the argument
list of the ISO_C_BINDING wrappers.  Updated the wrapper code call to
the original function to use keyword arguments, so that it will work
with various combinations of optional arguments and is able to skip
over the removed Fortran-only arguments.

Could consider updating so that keyword argument passing is only used
when needed (when there are Fortran-only arguments).

Some associated internal rework to create Argument.is_hidden to track
auxiliary string and array size arguments, which are no longer
considered in the Argument.fort_only function.

With these updates, the recently added Fortran-only case in the
optionals test now passes, as with all other tests.