Adding test for optional procedure pointers
There does not appear to be a straightforward way to wrap this in the
ISO_C_BINDING code, since the "present" status can't be represented in
the local variable used to store the Fortran function pointer.  For
now, the wrapper code always passes in a value for the dummy argument,
but the user code should make sure to check for null and treat that
the same as not present.  A warning has been added when optional
proccedure pointers are wrapped.