Add line continuations to long lines in generated wrapper code
Who would have known that free format has a line length limit?
Apparently it is 132 characters, and the ISO_C_BINDING wrapper code
can easily exceed this.  For now, adding continuations to break at 76
characters, and handling this for both the BIND(C) function definition
and the call into the original function.

Also added test case to the "syntax" tests, which failed to compile
prior to this update but now passes.