Minor tweak to handle array size declared as derived type component
Wrapper code converts it to assumed size ("*"), since the derived type
component is not available in the BIND(C) wrapper.  This could lead to
issues with multidimensional arrays that size based on derived type
components.  With proper identification, those cases could be handled
by passing a pointer and extra size arguments to the wrapper.

Added test to "arrays" to verify handling of derived type size argument.