Initial take at writing bind(c) wrappers to multiple modules
The motivation for this is that with large projects, there may be name
clashes in the Fortran API if all of the modules are USE'd at the same
time, which is what was previously done in the FortranISOWrappers
file.  Now when the --single-module option is not activated, bind(c)
wrappers for each individual module are written within a separate
wrapper module.  This avoids the potential name clashes, since all
modules aren't USE'd at once.

Some care is needed for what modules must be USE'd within each wrapper
module.  Currently it just USE's the original module.  However, this
won't work if the API includes derived types or abstract interfaces
that are not available in that module.

The --single-module option is now activated.  The default is to use
separate modules.  All existing tests pass with the new multi-module
approach.