Add necessary USE statements in wrapper modules
Make sure that wrapper modules USE additional modules when necessary
for derived types that are in the API.  Added new test "modules" that
demonstrates (a) a Fortran name clash if all modules are USE'd
together and (b) that the wrapper for module_b must also USE module_a.

The same issue could also arise with abstract interfaces, which will
need to be handled separately.