Fixes for logical arguments
Previous approach did not generate valid code for logical
arguments (function return values were OK).  Because the C_BOOL (1
byte) could not be passed to a regular Fortran Logical (4 bytes).
This differs from the other types because ISO_C_BINDING does not
provide a direct mapping for different kinds (in particular, the
default logical).  So have to add Fortran code to the wrapper to do
the conversion.  And for intent(inout) or intent(out), need to also
convert after the function call as well.

In addition, I have changed the C mapping type from int to bool (this
seemed to be necessary to make the full tests work on gcc).

These changes pass with new tests (will commit separately).  However,
I'm unsure how portable this is, in part because C++ bool is not
necessarily the same as C _Bool, which is technically what C_BOOL is
matching.