Handle wrapping of overloaded procedures
An INTERFACE block may be used to define a generic interface in
Fortran, in which case the actual procedure may not be public.  The
wrapper code needs to account for this when calling into the
function.  A new global variable is used to track interface
definitions in each modules, and these are checked when writing the
function call.  Overloading not yet done in the generated C++ API, but
believe that could be added.

Adding new test "overloading" that tests procedures that are only
public via a generic interface.