Fixes to support C_PTR wrapping
A few minor tweaks so that C_PTR arguments are wrapped.  Unset the
derived type flag so the argument is treated like a basic type.
Provide a C++ type map to void**.  Add handling for the VALUE
attribute.  If intent(in) is used without the VALUE attribute, the
wrapper code fixes up the C++ API to pass a value (here void*), just
like the other types.

Adding test case with TYPE(C_PTR).  Verifies intent(in), intent(out),
and use of the VALUE attribute.