Wrap abstract interfaces that use VALUE arguments
Minor cleanup so that these are wrapped correctly.  Adding new
function pointer test cases that demonstrate abstract interfaces with
VALUE arguments and BIND(C).

This was cherry picked from the iso_c_binding.  Unlike the original
commit, not allowing VALUE attribute for procedure pointers on this
branch, since the wrapping isn't set up to handle it (even though my
compiler (gfortran 5.4) won't allow that anyway).  However, this
reveals that the regular expression used to detect the VALUE attribute
is errenously flagging on the procedure interface part, which causes
the tests here to fail.  This should be addressed by updated the
regular expression.