Rework regular expressions for argument parsing
Add an "attributes" group to the fort_data_def regular expression to
capture attributes occuring beyond the data declaration.  Update code
so that specific attributes such as optional, value, allocatable,
pointer, etc. are only searched within the "attributes" text, as
opposed to over the entire line.  This helps avoid falsely picking up
an attribute if the text occurs within the interface definition of a
procedure pointer.

Some corresponding updates were needed to the procedure pointer
identification code within the DataType constructor.  It would
previously check for the pointer attribute, but that is no longer
included in the type string.  Now code after the DataType constructor
will make adjustements to the specification of the data type for
"procedure" arguments that also have the "pointer" attribute.

This resolves a test failure in "func_pointers" due to inadvertently
picking up the "value" attribute for a procedure pointer argument.