Add %init keyword to configuration file
The previously used "initialized" member is no longer added to proxy
classes by default.  Now it can be added using the %init keyword in
the configuration file.  The keyword specifies a regular expression
for initialization functions that must be called before the destructor
is called.

Updating manual to document %init