Add special handling of \par command in doxygen comments
When a \par command is encountered with a paragraph title, move all
subsequent generated comments to after the \param comments.  This is
intended to make it possible to have sections within the
comments (such as "Examples" or "Note" or "References") that appear
after the \param section.  See the updated example in
tests/comments/comments.f90.

One motivation for this is to support comment translation when
re-wrapping the generated C++ code for Python using swig.  The vision
is to target the numpy documentation style, in which additional
sections such as "Examples" follow the "Parameters" section.  Current
ongoing work on swig is adding support for doxygen comment translation
into sphinx format.  More testing will be needed to determine exactly
how best to make this work, but this is a start.